third_party/WebKit/LayoutTests/fullscreen/full-screen-remove-ancestor-during-transition.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

function runWithKeyDown(fn)
{
    document.addEventListener('keydown', function() { fn(); }, false);
    if (window.testRunner) {
        eventSender.keyDown('a');
    }
}

function init() {
    runWithKeyDown(goFullScreen);
}

function goFullScreen() {
    var iframe = document.getElementById('block1');
    var element = iframe.contentDocument.documentElement;
    setTimeout(function () {
        iframe.parentNode.removeChild(iframe);
        gc();
        setTimeout(function () {
            if (window.testRunner) {
                testRunner.notifyDone();
            }
        }, 0);
    }, 0);
    element.webkitRequestFullScreen();
    shouldNotBe("document.webkitFullscreenElement", "null");
}
</script>
<body onload="init()">
    <iframe allowfullscreen src="resources/inner.html" id="block1"></iframe>
    PASS
</body>
