third_party/WebKit/LayoutTests/fullscreen/full-screen-remove-ancestor-during-transition-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.webkitFullscreenElement is not null
PASS
